+++ CHANGELOG.md
@@ -6,11 +6,12 @@
 - **Breaking**: module: パラメーター取得系の関数を`Option`から`Result<T, GetParamError>`にする
   - Note: `#[aviutl2::module::functions]`を使っていて、かつ`#[direct]`を使用していない場合は追従は不要です。
 - **Breaking**: plugin: `EditInfo`のBPMグリッド系のフィールドを削除
+- **Breaking**: plugin: `EditSection`の単一BPMグリッド系のメソッドを削除
 - **Breaking**: alias: `TrackItem`を削除（正しいパースができなくなったため）
 
 ### デモプラグイン
 
-（なし）
+- metronome-plugin: 複数BPMグリッドに対応
 
 ## [0.35.0](https://github.com/sevenc-nanashi/aviutl2-rs/releases/tag/0.35.0)